load extra domains and main domain only once when starting console
