+++ apis/base_app.py
@@ -6,8 +6,8 @@ import collections
 from etcd import EtcdKeyNotFound
 from .specs import render_app_spec, AppType
 from lain_sdk.yaml.parser import ProcType, LainConf
-from commons.settings import PRIVATE_REGISTRY, DOMAIN, extra_domains
-from commons.utils import (
+from commons.settings import PRIVATE_REGISTRY
+from .utils import (
     normalize_meta_version,
     search_images_from_registry,
     get_meta_from_registry,
@@ -16,6 +16,7 @@ from commons.utils import (
     set_value_to_etcd,
     get_meta_version_from_tag,
     get_current_time,
+    get_domains,
 )
 from commons.miscs import (
     InvalidMetaVersion,
@@ -32,9 +33,6 @@ APP_RUNNING_STATE = {
     'DEPLOY' : 'deployed'
 }
 
-
-def get_domains():
-    return [DOMAIN] + extra_domains()
 
 class BaseApp:
